# .openapi/freelo-api-openapi.yaml
openapi: 3.0.3
info:
  title: Freelo API
  description: |
    API documentation for [Freelo.io](https://www.freelo.io) - Project Management Tool.
    
    ## Request Format
    All requests point to URL that begins with `https://api.freelo.io/v1/`.
    Each request must include a `User-Agent` HTTP header.
    
    ### Example: Projects List
    ```bash
    curl -u your@email.tld:API_KEY -H 'User-Agent: Your_App_Id (email@company.tld)' \
        https://api.freelo.io/v1/projects
    ```
    
    ## Response Format
    The API only supports JSON format in UTF-8 encoding.
    
    ## Authentication
    Authentication is done using HTTP Basic Authentication. Use your email as username and API key as password.
    You can find the API key in [settings](https://app.freelo.io/profil/nastaveni).
    
    ## Rate Limiting
    You can perform up to 25 requests per minute. If you exceed this limit, 429 Too Many Requests HTTP status code will be returned.
    You have to wait at least 60 seconds before sending another request.
    
    ## Currencies Format
    All currency amounts must be strings with exactly 2 decimal places multiplied by 100.
    
    Supported currencies: CZK, EUR, USD
    
    Examples:
    - `1,000.25` = `"100025"`
    - `1,000` = `"100000"`
    - `1` = `"100"`
    
    ## Pagination
    Pages are set via GET parameter `p` starting from 0.
    
    ## Ordering
    Use `order_by` and `order` (asc/desc) parameters where supported.

  version: "1.0.0"
  contact:
...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: Get all own active projects
      description: Collection of all own (active) projects including active tasklists
      operationId: getProjects
      parameters:
        - name: order_by
          in: query
          description: Order column
          schema:
            type: string
            enum: [name, date_add, date_edited_at]
            default: name
        - name: order
          in: query
          description: Order direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectWithTasklists'
    post:
      tags:
        - Projects
      summary: Create project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - currency_iso
              properties:
                name:
                  type: string
                  example: "Vegetable growing"
                currency_iso:
                  type: string
                  enum: [CZK, EUR, USD]
                project_owner_id:
                  type: integer
                  description: ID of user assigned as owner
      responses:
        '200':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectBasic'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Use your email as username and API key as password
...
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        company:
          type: string
        company_id:
          type: string
        company_tax_id:
          type: string
        street:
          type: string
        town:
          type: string
        zip:
          type: string
    TasklistBasic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    ProjectBasic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    ProjectWithTasklists:
      allOf:
        - $ref: '#/components/schemas/ProjectBasic'
        - type: object
          properties:
            date_add:
              type: string
              format: date-time
            date_edited_at:
              type: string
              format: date-time
            tasklists:
              type: array
              items:
                $ref: '#/components/schemas/TasklistBasic'
            client:
              $ref: '#/components/schemas/Client'
...